+++ reiwa/tsubasa-overflow/main.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 
 int main() {
-  int  a, n, t;
+  short  a, n, t;
   scanf("%d", &t);
   for(int tt = 0; tt < t; ++tt) {
     scanf("%d", &n);
     a = 0;
     for(int i = 0; i < n - 1; ++i) {
-      int x;
+      short x;
       scanf("%d", &x);
       a += x;
     }
